feat: content to be saved does not need to be base64 encoded beforehand
The library now manages base64 encoding.

Note 1: `type-fest` is in the `pkg.dependencies` so that it is also made
available to consumers of this library

Note 2: `type-fest`4.41.0 requires typescript >= 5.1 according to its documentation in npm.
Future versions will require even higher typescript versions (as can be seen already
in the documentation in github - not published yet in npm).

Note 3: `@types/node` is NOT in `pkg.dependencies` but in `pkg.devDependencies` instead
because it is assumed the consumer will have it anyway in its own list of dependencies.